+++ biu/CustomizeUI/BLSignupView.m
@@ -362,7 +362,7 @@
 //    NSLog(@"code: %@", _code);
     [[BLHTTPClient sharedBLHTTPClient] passcode:_code phoneNumber:_tfPhoneNumber.text success:^(NSURLSessionDataTask *task, id responseObject) {
         _secondLeftToResend = 60;
-        _lbSecondLeft.text = [NSString stringWithFormat:@"%ld", _secondLeftToResend];
+        _lbSecondLeft.text = [NSString stringWithFormat:@"%ld", (long)_secondLeftToResend];
         _btnGetCode.enabled = NO;
         [self showSecondToResend];
         _timer = [NSTimer scheduledTimerWithTimeInterval:1 target:self selector:@selector(timerfired) userInfo:nil repeats:YES];
